requirements.txt
streamlit
groq
pymupdf
python-dotenv
fpdf2
